# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Practise Project",
        "version": "0.1"
    },
    "paths": {
        "/api/v1/post": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Posts",
                "operationId": "eaee9fa5ab200c8ffe5b90d86de17bf0",
                "responses": {
                    "200": {
                        "description": "Return all Posts"
                    },
                    "403": {
                        "description": "this action id Unauthorized"
                    }
                }
            }
        },
        "/api/v1/post/{post}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Show Post",
                "operationId": "baf761218df3533d90892fd1a7c6ea97",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "You must send the post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return a specific post"
                    },
                    "403": {
                        "description": "this action id Unauthorized"
                    }
                }
            }
        },
        "/api/v1/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Users",
                "operationId": "cd657b847bd27cccecf4b00564dbf99f",
                "responses": {
                    "200": {
                        "description": "Return all Users"
                    },
                    "403": {
                        "description": "this action id Unauthorized"
                    }
                }
            }
        },
        "/api/v1/user/{user}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Show User",
                "operationId": "2d5c6f3e4267c1384fe834fb965a101d",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "You must send the user ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return a specific user"
                    },
                    "403": {
                        "description": "this action id Unauthorized"
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Authentication-system"
                ],
                "summary": "Login",
                "operationId": "780c560985ffaa396d3a61c81b0a814f",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login was successful"
                    },
                    "401": {
                        "description": "Login is not possible for you"
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "get": {
                "tags": [
                    "Authentication-system"
                ],
                "summary": "Logout",
                "responses": {
                    "200": {
                        "description": "Logout was successful"
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Authentication-system"
                ],
                "summary": "Register",
                "operationId": "3645743e46c9d8f5ce43d34260eef32e",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registration was successful"
                    }
                }
            }
        }
    }
}